removed debug outputs

+++ bundlelibs.rb
@@ -52,18 +52,22 @@ end
 
 def is_lib_need_to_bundle(lib_path)
 	
+	# relative path
 	if not Regexp.new("^/") =~ lib_path
 		return false
 	end
 	
+	#system libs 
 	if Regexp.new("^/usr/lib") =~ lib_path
 		return false
 	end
 	
+	#system libs
 	if Regexp.new("^/System/Library/Frameworks") =~ lib_path
 		return false
 	end
 	
+	# nothing to be done
 	if /^@executable_path/ =~ lib_path
 		return false
 	end
@@ -73,10 +77,6 @@ def is_lib_need_to_bundle(lib_path)
 end
 
 def change_install_name_lib(binary, before, after)
-	puts "binary: #{binary}"
-	puts "before: #{before}"
-	puts "after: #{after}"
-	
 	`install_name_tool -change #{before} #{after} #{binary}`
 end
 
@@ -112,7 +112,14 @@ end
 def bundle_dylib(src, dst_dir, executable_path)
 	
 	dst = get_dst_path(src, dst_dir)
-	p dst
+	
+	if not Pathname.new(src).exist?
+		raise "dylib not found"
+	end
+	
+	if Pathname.new(dst).exist?
+		raise "dylib already exists in destination"
+	end
 	
 	FileUtils.cp(src, dst)
 	FileUtils.chmod(0644, dst)
@@ -123,7 +130,14 @@ end
 def bundle_framework(src, dst_dir, executable_path)
 	
 	dst = get_dst_path_framework(src, dst_dir)
-	p dst
+	
+	if not Pathname.new(get_framework_path(src)).exist?
+		raise "framework not found"
+	end
+	
+	if Pathname.new(get_framework_path(dst)).exist?
+		raise "framework already exists"
+	end
 	
 	FileUtils.cp_r(get_framework_path(src), dst_dir)
 	FileUtils.chmod(0644, dst)
@@ -133,11 +147,11 @@ end
 
 $already_bundled_list = Array.new
 
-def is_already_included(dst)
+def is_already_included(name)
 	
 	$already_bundled_list.each do |bundled|
 		
-		if bundled == dst
+		if bundled == name
 			return true
 		end
 		
@@ -147,11 +161,13 @@ def is_already_included(dst)
 	
 end
 
-def add_already_included_lib(dst)
+def add_already_included_lib(name)
 	
-	$already_bundled_list << dst
+	$already_bundled_list << name
 	
 end
+
+$bundled_lib_src_names = []
 
 def bundle_used_libs(binary, bundle_dir, executable_path)
 	
@@ -160,6 +176,8 @@ def bundle_used_libs(binary, bundle_dir, executable_path)
 		if is_lib_need_to_bundle(name)
 			
 			if is_dylib(name)
+				
+				$bundled_lib_src_names = $bundled_lib_src_names | [name]
 				
 				dst_name = get_dst_path(name, bundle_dir)
 				change_install_name_lib(binary, name, get_relative_install_name(dst_name, executable_path))
@@ -175,6 +193,8 @@ def bundle_used_libs(binary, bundle_dir, executable_path)
 			end
 			
 			if is_framework(name)
+				
+				$bundled_lib_src_names = $bundled_lib_src_names | [name]
 				
 				dst_name = get_dst_path_framework(name, bundle_dir)
 				change_install_name_lib(binary, name, get_relative_install_name(dst_name, executable_path))
@@ -202,7 +222,6 @@ def main
 		
 		opts.banner = "Usage: bundlelibs.rb -x [binary filepath] -d [destination directory path]"
 		opts.on("-x BINARY_PATH") do |path|
-			p path
 			app_pathname = Pathname.new(path)
 		end 
 		opts.on("-d DESTINATION_PATH") do |path|
@@ -224,11 +243,22 @@ def main
 		app_pathname = app_pathname.realpath
 	end
 	
+	if not app_pathname.exist?
+		raise "binary not found"
+	end
+	
 	if bundle_pathname.relative?
 		bundle_pathname = bundle_pathname.realpath
 	end
 	
+	bundle_pathname.mkpath
+	
 	bundle_used_libs(app_pathname.to_s, bundle_pathname.to_s, app_pathname.parent.to_s)
+	
+	puts "Bundled libs:"
+	$bundled_lib_src_names.each do |name|
+		puts name
+	end
 	
 end
